requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
django-filter>=23.0
markdown>=3.5.0
drf-yasg>=1.21.0
djangorestframework-simplejwt>=5.2.2
django-debug-toolbar
gunicorn